# internal/docs/swagger.yaml
definitions:
...
  api.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/api.UserResponse'
      workspace:
        $ref: '#/definitions/api.WorkspaceResponse'
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  api.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  api.ValidationErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
...
  api.WorkspaceResponse:
    properties:
      created_at:
        type: string
      download_log_retention_days:
        type: integer
      event_log_retention_days:
        type: integer
      exif_keep:
        type: boolean
      exif_keep_gps:
        type: boolean
      icon_asset_id:
        type: string
      icon_version_id:
        type: string
      id:
        type: string
      ingest_token:
        type: string
      name:
        type: string
      updated_at:
        type: string
      version_retention_count:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates the user with email and password. Returns a JWT token
        in the response body and sets an <code>auth_token</code> httpOnly cookie (7-day
        expiry). The workspace embedded in the response is the user's first workspace
        — use <code>POST /api/v1/workspace/switch</code> to change it.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
      summary: Log in
      tags:
      - Auth
...
  /auth/refresh:
    post:
      description: Reissues a fresh JWT with a new 7-day expiry for the authenticated
        user and workspace. The new token is returned in the response body and also
        set as an updated <code>auth_token</code> cookie. Call this endpoint before
        the current token expires to maintain a session without re-login.
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh auth token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with name, email, and password. A default
        workspace is automatically created for the new user. The response includes
        a JWT auth token set as an <code>auth_token</code> httpOnly cookie and also
        returned in the response body.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/api.ValidationErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
swagger: "2.0"